+++ injectool/__init__.py
@@ -1,6 +1,6 @@
 """Dependency injection tool"""
 
-__version__ = '1.1.1'
+__version__ = '1.1.2'
 
 from .core import DependencyError, Container, SingletonResolver, FunctionResolver
 from .core import make_default, add_resolver, add_singleton, add_function_resolver, resolve